requirements.txt
matplotlib==3.9.2
PyQt6==6.7.1
PyQt6-Qt6==6.7.3
PyQt6_sip==13.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3